# pos-invoice/openapi.yaml
openapi: 3.1.0
info:
  title: POS Invoice API
  description: API documentation for invoice and billing rules management
  contact:
    name: Durion Support Services
    email: platform@durionpos.org
  version: v1
servers:
- url: http://localhost:8086
  description: Generated server url
...
paths:
...
  /v1/invoices:
    post:
      tags:
      - Invoice
      summary: Create invoice
      description: Create invoice draft from completed workorder data
      operationId: createInvoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreationRequest'
        required: true
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceGenerationResponse'
      security:
      - bearerAuth:
        - invoice:manage
      x-required-permissions:
      - invoice:manage
...
  /v1/invoices/{invoiceId}/payments:
    post:
      tags:
      - Payment
      summary: Initiate card payment
      description: Initiate a SALE_CAPTURE or AUTH_ONLY card payment against an invoice
      operationId: initiatePayment
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiatePaymentRequest'
        required: true
      responses:
        '201':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentResponse'
        '400':
          description: Invalid or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentResponse'
        '422':
          description: Payment method declined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentResponse'
        '503':
          description: Payment gateway unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentResponse'
      security:
      - bearerAuth: []
      x-required-permissions:
      - AUTHENTICATED
...
  /v1/invoices/{invoiceId}/payments/{paymentId}/capture:
    post:
      tags:
      - Payment
      summary: Capture authorized payment hold
      description: Capture all or part of a previously authorized invoice payment hold
      operationId: capturePayment
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: paymentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureAmountRequest'
        required: true
      responses:
        '200':
          description: Payment captured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentResponse'
        '404':
          description: Payment intent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentResponse'
      security:
      - bearerAuth: []
      x-required-permissions:
      - AUTHENTICATED
...
  /v1/invoices/{invoiceId}/artifacts/{artifactRefId}/download:
    get:
      tags:
      - invoice-artifact-download-controller
      summary: Download an invoice artifact as PDF using a signed download token
      description: Streams the artifact as application/pdf. Public endpoint authorized solely by the signed download token (query parameter); intended for browser direct-download links.
      operationId: download
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: artifactRefId
        in: path
        required: true
        schema:
          type: string
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: PDF returned
          content:
            application/json:
              schema:
                type: string
                format: byte
        '403':
          description: Missing, invalid, or expired token
          content:
            application/json:
              schema:
                type: string
                format: byte
        '404':
          description: Invoice or artifact not found
          content:
            application/json:
              schema:
                type: string
                format: byte
...
components:
  schemas:
...
    InvoiceCreationRequest:
      type: object
      description: Internal service payload for creating invoice drafts.
      properties:
        workorderId:
          type: string
          format: uuid
          description: Workorder identifier backing the invoice draft.
          example: 550e8400-e29b-41d4-a716-446655440000
        estimateId:
          type: string
          format: uuid
          description: Source estimate identifier, when generated from estimate.
          example: 550e8400-e29b-41d4-a716-446655440001
        approvalId:
          type: string
          format: uuid
          description: Approval identifier used for billing authorization.
          example: 550e8400-e29b-41d4-a716-446655440002
        locationId:
          type: string
          format: uuid
          description: Shop location where the sale is made; used to resolve the tax jurisdiction address.
          example: 01960003-0000-7000-8000-000000000001
        idempotencyKey:
          type: string
          description: Idempotency key to prevent duplicate invoice creation.
          example: inv-create-wo-1234
        lineItems:
          type: array
          description: Line items to include on the invoice.
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
    InvoiceLineItem:
      type: object
      description: Invoice line item details.
      properties:
        description:
          type: string
          description: Line item description.
          example: Brake pad replacement
        quantity:
          type: number
          description: Quantity billed.
          example: 2.0
        unitPrice:
          type: number
          description: Unit price amount.
          example: 45.0
        amount:
          type: number
          description: Line amount (quantity x unit price).
          example: 90.0
      required:
      - amount
      - description
      - quantity
      - unitPrice
    InvoiceGenerationResponse:
      type: object
      description: Response payload returned after invoice generation.
      properties:
        invoiceId:
          type: string
          format: uuid
          description: Generated invoice identifier.
          example: 550e8400-e29b-41d4-a716-446655440000
        status:
          type: string
          description: Invoice status.
          example: DRAFT
        workorderId:
          type: string
          format: uuid
          description: Source workorder identifier.
          example: 550e8400-e29b-41d4-a716-446655440001
        estimateId:
          type: string
          format: uuid
          description: Source estimate identifier.
          example: 550e8400-e29b-41d4-a716-446655440002
        approvalId:
          type: string
          format: uuid
          description: Approval identifier used for generation.
          example: 550e8400-e29b-41d4-a716-446655440003
        subtotal:
          type: number
          description: Subtotal amount before tax.
          example: 120.0
        taxAmount:
          type: number
          description: Tax amount.
          example: 9.6
        totalAmount:
          type: number
          description: Total amount after tax.
          example: 129.6
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (UTC).
          example: 2026-02-27 12:00:00+00:00
      required:
      - createdAt
      - invoiceId
      - status
      - subtotal
      - taxAmount
      - totalAmount
      - workorderId
...
    InitiatePaymentRequest:
      type: object
      description: Request to initiate a card payment
      properties:
        paymentFlow:
          type: string
          description: 'Payment flow: SALE_CAPTURE (default one-step) or AUTH_ONLY (two-step)'
          enum:
          - SALE_CAPTURE
          - AUTH_ONLY
          example: SALE_CAPTURE
        amount:
          type: number
          description: Payment amount (must be positive)
          example: 149.99
        idempotencyKey:
          type: string
          description: Client-provided idempotency key for safe retry
          example: idem-01960003-0000-7000-8000-000000000010
          minLength: 1
        paymentToken:
          type: string
          description: Tokenised card reference (no PAN)
          example: tok_1NXyZ2eZvKYlo2Cabc12345
          minLength: 1
      required:
      - amount
      - idempotencyKey
      - paymentFlow
      - paymentToken
    InitiatePaymentResponse:
      type: object
      description: Payment intent result
      properties:
        paymentIntentId:
          type: string
          format: uuid
          description: Payment intent identifier
          example: 01960003-0000-7000-8000-000000000020
        status:
          type: string
          description: Current lifecycle status of the payment intent
          enum:
          - PENDING
          - AUTHORIZED
          - CAPTURED
          - CAPTURE_FAILED
          - VOIDED
          - EXPIRED
          example: AUTHORIZED
        authorizedAmount:
          type: number
          description: Amount authorized on the card
          example: 149.99
        capturedAmount:
          type: number
          description: Amount captured (funds moved)
          example: 149.99
        voidedRemainderAmount:
          type: number
          description: Remainder voided after partial capture
          example: 0.0
        gatewayProvider:
          type: string
          description: Gateway provider identifier (e.g., stripe)
          example: stripe
        gatewayResponse:
          type: string
          description: Raw gateway response JSON for audit
          example:
            id: pi_3NXyZ2eZvKYlo2Ca
            status: requires_capture
      required:
      - paymentIntentId
      - status
...
    CaptureAmountRequest:
      type: object
      description: Request to capture all or part of an authorized payment hold
      properties:
        amount:
          type: number
          description: Amount to capture from the authorized hold
          example: 89.99
        captureIdempotencyKey:
          type: string
          description: Idempotency key ensuring the capture is processed at most once
          example: capture-550e8400-1
          minLength: 1
      required:
      - amount
      - captureIdempotencyKey
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT